Add bid war option to require use of the !bid command.
When this option is set, the bot will ignore resub messages, bit messages, etc. Points will only be assigned when a user uses the !bid command in chat.

This is useful when one of the bid war commands is a common word that chatters are likely to say by accident.